web: behaviour: filter based on `DashboardView`
concourse/concourse#5640

Toggling the switch on now allows archived pipelines to be displayed. I
initially implemented this as a pipeline filter, where each
`DashboardView` had a list of associated `Filter`s, which felt like a
good way to do things considering the intended future use cases of
`DashboardView`. However, the current behaviour is to not render a team
section (i.e. `Group`) if a `Filter` does not match.

So, if `ViewNonArchivedPipelines` had a `Filter` associated, teams with
no non-archived pipelines would not be rendered at all, which is not
desired behaviour (we want to display the "no pipelines" card instead as
we currently do).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>