fly: behaviour: allow pasting long tokens into login
concourse/concourse#5717

On OSX, there's some magic constant of 1024 characters, which is the
length of some TTY buffer. Not sure of the specifics, but essentially
you weren't able to paste tokens longer than 1024 characters in `fly
login` for the "or enter token manually" option.

Switching to a buffered IO reader seems to fix this (I guess it's
flushing the TTY buffer and moving it into a larger buffer in code?) The
one downside is that you can no longer e.g. backspace to delete content,
but that's probably okay given that you aren't going to be typing out
your token by hand.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>